feat: move ReplicateUserFromTemplate logic to UseReplicateUserFromTemplateReturn. Use new User entity. Merge API call useEffects into one.
